requirements.txt
streamlit
pandas
numpy
faiss-cpu
openai
transformers
sentence-transformers
sentencepiece
python-dotenv
